requirements.txt
streamlit
newspaper3k
vaderSentiment
spacy
bs4
pandas
matplotlib
